backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-rest-framework==0.1.0
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uritemplate==4.1.1
